feat: docker init

+++ frontend/src/shared/api/uploadthing/actions.ts
@@ -5,6 +5,7 @@ import { utapi } from './utapi'
 
 export const utapiDeleteFiles = async (file: IAsset) => {
 	"use server"
+	if (!file.key) return
 	const result = await utapi.deleteFiles(file.key)
 	if (result.success) {
 		return {
